requirements.txt
openai-whisper
torch
librosa
soundfile
numpy
scipy